# swagger.yaml
swagger: '2.0'
info:
  description: How to use the Boastable server.
  version: 1.0.0
  title: Boastable API
  contact:
    email: ryan@fueledonbacon.com
host: localhost:3000
basePath: /
...
schemes:
  - http
paths:
...
  /question/list:
    get:
      tags:
        - questions
      summary: List questions
      description: 'Returns all questions from customers a user is allowed to see or query. Client results will be filtered to only return questions they created'
      operationId: listAnswers
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: query
          in: query
          description: Any key from the customer document
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Answer'
        '400':
          description: Invalid username supplied
        '404':
          description: Action not found
  /question:
    post:
      tags:
        - questions
      summary: Create question
      description: Questions created by clients to get feedback from customers
      operationId: createQuestion
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - in: body
          name: body
          description: Created user object
          required: true
          schema:
            $ref: '#/definitions/Question'
      responses:
        default:
          description: successful operation
    get:
      tags:
        - questions
      summary: Get question
      description: ''
      operationId: getQuestion
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: username
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
    put:
      tags:
        - questions
      summary: Updated question
      description: Update the text of a question
      operationId: updateQuestion
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: question that needs to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated user object
          required: true
          schema:
            $ref: '#/definitions/Question'
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: User not found
    delete:
      tags:
        - questions
      summary: Delete question
      description: Only an admin can do this
      operationId: deleteQuestion
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: An action that needs to be deleted
          required: true
          type: string
      responses:
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Action not found
...
  /customer/list:
    get:
      tags:
        - customers
      summary: List customers
      description: 'Returns all customers user is allowed to see or query. Client results will be filtered to only return Customers they created'
      operationId: listCustomers
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: query
          in: query
          description: Any key from the customer document
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Customer'
        '400':
          description: Invalid username supplied
        '404':
          description: Action not found
  /customer:
    post:
      tags:
        - customers
      summary: Create customer
      description: Clients create their own customers
      operationId: createCustomer
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - in: body
          name: body
          description: Created user object
          required: true
          schema:
            $ref: '#/definitions/Customer'
      responses:
        default:
          description: successful operation
    get:
      tags:
        - customers
      summary: Get customer
      description: 'Customers have a phone number associated with their document. Clients may only request phone numbers for customers they added'
      operationId: getCustomer
      produces:
        - application/json
      parameters:
        - name: id
          in: query
          description: The ID of the customer
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Customer'
        '400':
          description: Invalid username supplied
        '404':
          description: Action not found
    put:
      tags:
        - customers
      summary: Update customer data
      description: This can only be done by the logged in user.
      operationId: updateCustomer
      produces:
        - application/json
      parameters:
        - name: id
          in: query
          description: ID of the customer that needs to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated customer object
          required: true
          schema:
            $ref: '#/definitions/Customer'
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: Action not found
    delete:
      tags:
        - customers
      summary: Delete customer
      description: Only an admin can do this
      operationId: deleteCustomer
      produces:
        - application/json
      parameters:
        - name: id
          in: query
          description: A customer that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Action not found
definitions:
...
  Customer:
    type: object
    required:
      - email
      - phone
    properties:
      _id:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      auth:
        type: string
        description: ObjectId, the client authorized to see this customer's information
        example: 5a2cfe672ebedf740dac80c8
      email:
        type: string
        example: example@user.com
      phone:
        type: string
        description: An E.164 phone number
        example: '+15616235746'
...
  Question:
    type: object
    required:
      - user
      - question
    properties:
      _id:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      auth:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      question:
        type: string
        description: The question being asked
        example: Was service at our location good?
  Answer:
    type: object
    required:
      - customer
      - user
      - answer
      - question
    properties:
      _id:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      customer:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      auth:
        type: string
        description: ObjectId, the client authorized to see this customer's information
        example: 5a2cfe672ebedf740dac80c8
      question:
        type: string
        description: ObjectId
        example: 5a2cfe672ebedf740dac80c8
      answer:
        type: boolean
        description: A true false answer
        example: true
...
  User:
    type: object
    required:
      - auth
    properties:
      _id:
        type: string
        description: Mongo document ObjectId
        example: 5a2cfe672ebedf740dac80c8
      auth:
        type: string
        description: ObjectId reference to an Auth document
        example: 5a2cfe672ebedf740dac80c8
      first_name:
        type: string
        example: Bilbro
      last_name:
        type: string
        example: Baggins
      phone:
        type: string
        description: An E.164 formatted string
        example: +15615236666
...